fix: refactoring folder
